+++ hooks/Auth/useAuthRedirect.ts
@@ -1,7 +1,5 @@
 import { useRouter } from 'next/router';
 import { useEffect } from 'react';
-import { usePopup } from './usePopup';
-import useLoginState from './useLoginState';
 
 export default function useAuthRedirect() {
   const router = useRouter();
