+++ hooks/myActivity/useMyActivityList.ts
@@ -2,7 +2,7 @@ import { useInfiniteQuery } from '@tanstack/react-query';
 import { getMyActivityList } from '@/pages/api/myActivities/apimyActivities';
 import { getMyActivityListResponse } from '@/pages/api/myActivities/apimyActivities.types';
 import { useMemo } from 'react';
-import useLoginState from '../useLoginState';
+import useLoginState from '../Auth/useLoginState';
 
 const INITIAL_SIZE = 4;
 const REFETCH_SIZE = 1;
