+++ pages/login.tsx
@@ -5,9 +5,9 @@ import { loginFormValues } from '@/components/AuthInputBox/AuthInputBox.types';
 import { loginValidation } from '@/components/AuthInputBox/validation';
 import Link from 'next/link';
 import { PrimaryButton } from '@/components/Button/Button';
-import useLogin from '@/hooks/useLogin';
+import useLogin from '@/hooks/Auth/useLogin';
 import { useEffect, useMemo, useState } from 'react';
-import useLoginState from '@/hooks/useLoginState';
+import useLoginState from '@/hooks/Auth/useLoginState';
 import { useRouter } from 'next/router';
 import { GetServerSideProps } from 'next';
 import useEnterSubmit from '@/hooks/useEnterSubmit';
