+++ pages/mypage.tsx
@@ -3,8 +3,8 @@ import { SSRMetaProps } from '@/components/MetaData/MetaData.type';
 import MyPageInput from '@/components/MyPageInput/MyPageInput';
 import SidenNavigation from '@/components/SideNavigation/SideNavigation';
 import SidenNavigationMobile from '@/components/SideNavigation/SideNavigationMobile';
-import useAuthRedirect from '@/hooks/useAuthRedirect';
-import useLoginState from '@/hooks/useLoginState';
+import useAuthRedirect from '@/hooks/Auth/useAuthRedirect';
+import useLoginState from '@/hooks/Auth/useLoginState';
 import { useSideNavigation } from '@/hooks/useSideNavigation';
 import { GetServerSideProps } from 'next';
 import { useRouter } from 'next/router';
